Added pod 2 pod connectivity tests
